refactor: Refactored unet models
