+++ holocron/models/segmentation/unet.py
@@ -13,22 +13,13 @@ from ...nn.init import init_module
 from ..utils import conv_sequence, load_pretrained_params
 
 
-__all__ = ['UNet', 'unet', 'UNetp', 'unetp', 'UNetpp', 'unetpp', 'UNet3p', 'unet3p']
+__all__ = ['UNet', 'unet']
 
 
 default_cfgs: Dict[str, Dict[str, Any]] = {
     'unet': {'arch': 'UNet',
              'layout': [64, 128, 256, 512, 1024],
              'url': None},
-    'unetp': {'arch': 'UNetp',
-              'layout': [64, 128, 256, 512, 1024],
-              'url': None},
-    'unetpp': {'arch': 'UNetpp',
-               'layout': [64, 128, 256, 512, 1024],
-               'url': None},
-    'unet3p': {'arch': 'UNet3p',
-               'layout': [64, 128, 256, 512, 1024],
-               'url': None}
 }
 
 
@@ -56,7 +47,6 @@ class UpPath(nn.Module):
         self,
         in_chan: int,
         out_chan: int,
-        num_skips: int = 1,
         bilinear_upsampling: bool = True,
         padding: int = 0,
         act_layer: Optional[nn.Module] = None,
@@ -78,14 +68,11 @@ class UpPath(nn.Module):
                                    *conv_sequence(out_chan, out_chan,
                                                   act_layer, norm_layer, drop_layer, conv_layer,
                                                   kernel_size=3, padding=padding))
-        self.num_skips = num_skips
 
     def forward(self, downfeats: Union[Tensor, List[Tensor]], upfeat: Tensor) -> Tensor:
 
         if not isinstance(downfeats, list):
             downfeats = [downfeats]
-        if len(downfeats) != self.num_skips:
-            raise ValueError(f"Expected {self.num_skips} encoding feats, received {len(downfeats)}")
         # Upsample expansive features
         _upfeat = self.upsample(upfeat)
         # Crop contracting path features
@@ -146,7 +133,7 @@ class UNet(nn.Module):
         self.decoders = nn.ModuleList([])
         _layout = [chan // 2 if bilinear_upsampling else chan for chan in layout[::-1][1:-1]] + [layout[0]]
         for in_chan, out_chan in zip(layout[::-1][:-1], _layout):
-            self.decoders.append(UpPath(in_chan, out_chan, 1, bilinear_upsampling, int(same_padding),
+            self.decoders.append(UpPath(in_chan, out_chan, bilinear_upsampling, int(same_padding),
                                         act_layer, norm_layer, drop_layer, conv_layer))
 
         # Classifier
@@ -168,253 +155,6 @@ class UNet(nn.Module):
 
         # Classifier
         x = self.classifier(x)
-        return x
-
-
-class UNetp(nn.Module):
-    """Implements a UNet+ architecture
-
-    Args:
-        layout (list<int>): number of channels after each contracting block
-        in_channels (int, optional): number of channels in the input tensor
-        num_classes (int, optional): number of output classes
-        act_layer: activation layer
-        norm_layer: normalization layer
-        drop_layer: dropout layer
-        conv_layer: convolutional layer
-    """
-    def __init__(
-        self,
-        layout: List[int],
-        in_channels: int = 3,
-        num_classes: int = 10,
-        act_layer: Optional[nn.Module] = None,
-        norm_layer: Optional[Callable[[int], nn.Module]] = None,
-        drop_layer: Optional[Callable[..., nn.Module]] = None,
-        conv_layer: Optional[Callable[..., nn.Module]] = None,
-    ) -> None:
-        super().__init__()
-
-        if act_layer is None:
-            act_layer = nn.ReLU(inplace=True)
-
-        # Contracting path
-        self.encoders = nn.ModuleList([])
-        _layout = [in_channels] + layout
-        _pool = False
-        for in_chan, out_chan in zip(_layout[:-1], _layout[1:]):
-            self.encoders.append(down_path(in_chan, out_chan, _pool, 1,
-                                           act_layer, norm_layer, drop_layer, conv_layer))
-            _pool = True
-
-        # Expansive path
-        self.decoders = nn.ModuleList([])
-        for next_chan, row_chan, num_cells in zip(layout[1:], layout[:-1], range(len(_layout) - 1, 0, -1)):
-            self.decoders.append(nn.ModuleList([
-                UpPath(next_chan + row_chan, row_chan, 1, True, 1,
-                       act_layer, norm_layer, drop_layer, conv_layer)
-                for _ in range(num_cells + 1)
-            ]))
-
-        # Classifier
-        self.classifier = nn.Conv2d(layout[0], num_classes, 1)
-
-        init_module(self, 'relu')
-
-    def forward(self, x: Tensor) -> Tensor:
-
-        xs: List[Tensor] = []
-        # Contracting path
-        for encoder in self.encoders:
-            xs.append(encoder(xs[-1] if len(xs) > 0 else x))
-
-        # Nested expansive path
-        for j in range(len(self.decoders)):
-            for i in range(len(self.decoders) - j):
-                xs[i] = self.decoders[i][j](xs[i], xs[i + 1] if (i + 1) < (len(self.decoders) - j) else xs.pop())
-
-        return self.classifier(xs.pop())
-
-
-class UNetpp(nn.Module):
-    """Implements a UNet++ architecture
-
-    Args:
-        layout (list<int>): number of channels after each contracting block
-        in_channels (int, optional): number of channels in the input tensor
-        num_classes (int, optional): number of output classes
-        act_layer: activation layer
-        norm_layer: normalization layer
-        drop_layer: dropout layer
-        conv_layer: convolutional layer
-    """
-    def __init__(
-        self,
-        layout: List[int],
-        in_channels: int = 3,
-        num_classes: int = 10,
-        act_layer: Optional[nn.Module] = None,
-        norm_layer: Optional[Callable[[int], nn.Module]] = None,
-        drop_layer: Optional[Callable[..., nn.Module]] = None,
-        conv_layer: Optional[Callable[..., nn.Module]] = None
-    ) -> None:
-        super().__init__()
-
-        if act_layer is None:
-            act_layer = nn.ReLU(inplace=True)
-
-        # Contracting path
-        self.encoders = nn.ModuleList([])
-        _layout = [in_channels] + layout
-        _pool = False
-        for in_chan, out_chan in zip(_layout[:-1], _layout[1:]):
-            self.encoders.append(down_path(in_chan, out_chan, _pool, 1, act_layer, norm_layer, drop_layer, conv_layer))
-            _pool = True
-
-        # Expansive path
-        self.decoders = nn.ModuleList([])
-        for next_chan, row_chan, num_cells in zip(layout[1:], layout[:-1], range(len(_layout) - 1, 0, -1)):
-            self.decoders.append(nn.ModuleList([
-                UpPath(next_chan + num_skips * row_chan, row_chan, num_skips, True, 1,
-                       act_layer, norm_layer, drop_layer, conv_layer)
-                for num_skips in range(1, num_cells + 2)
-            ]))
-
-        # Classifier
-        self.classifier = nn.Conv2d(layout[0], num_classes, 1)
-
-        init_module(self, 'relu')
-
-    def forward(self, x: Tensor) -> Tensor:
-
-        xs: List[List[Tensor]] = []
-        # Contracting path
-        for encoder in self.encoders:
-            xs.append([encoder(xs[-1][0] if len(xs) > 0 else x)])
-
-        # Nested expansive path
-        for j in range(len(self.decoders)):
-            for i in range(len(self.decoders) - j):
-                xs[i].append(self.decoders[i][j](
-                    xs[i][:j + 1],
-                    xs[i + 1][j] if (i + 1) < (len(self.decoders) - j) else xs.pop()[-1]
-                ))
-
-        # Classifier
-        x = self.classifier(xs.pop()[-1])
-        return x
-
-
-class FSAggreg(nn.Module):
-    def __init__(
-        self,
-        e_chans: List[int],
-        skip_chan: int,
-        d_chans: List[int],
-        act_layer: Optional[nn.Module] = None,
-        norm_layer: Optional[Callable[[int], nn.Module]] = None,
-        drop_layer: Optional[Callable[..., nn.Module]] = None,
-        conv_layer: Optional[Callable[..., nn.Module]] = None
-    ) -> None:
-
-        super().__init__()
-
-        # Check stem conv channels
-        base_chan = e_chans[0] if len(e_chans) > 0 else skip_chan
-        # Get UNet depth
-        depth = len(e_chans) + 1 + len(d_chans)
-        # Downsample = max pooling + conv for channel reduction
-        self.downsamples = nn.ModuleList([nn.Sequential(nn.MaxPool2d(2 ** (len(e_chans) - idx)),
-                                                        nn.Conv2d(e_chan, base_chan, 3, padding=1))
-                                          for idx, e_chan in enumerate(e_chans)])
-        self.skip = nn.Conv2d(skip_chan, base_chan, 3, padding=1) if len(e_chans) > 0 else nn.Identity()
-        # Upsample = bilinear interpolation + conv for channel reduction
-        self.upsamples = nn.ModuleList([nn.Sequential(nn.Upsample(scale_factor=2 ** (idx + 1),
-                                                                  mode='bilinear', align_corners=True),
-                                                      nn.Conv2d(d_chan, base_chan, 3, padding=1))
-                                        for idx, d_chan in enumerate(d_chans)])
-
-        self.block = nn.Sequential(*conv_sequence(depth * base_chan, depth * base_chan,
-                                                  act_layer, norm_layer, drop_layer, conv_layer,
-                                                  kernel_size=3, padding=1))
-
-    def forward(self, downfeats: List[Tensor], feat: Tensor, upfeats: List[Tensor]):
-
-        if len(downfeats) != len(self.downsamples) or len(upfeats) != len(self.upsamples):
-            raise ValueError(f"Expected {len(self.downsamples)} encoding & {len(self.upsamples)} decoding features, "
-                             f"received: {len(downfeats)} & {len(upfeats)}")
-
-        # Concatenate full-scale features
-        x = torch.cat((*[downsample(downfeat) for downsample, downfeat in zip(self.downsamples, downfeats)],
-                       self.skip(feat),
-                       *[upsample(upfeat) for upsample, upfeat in zip(self.upsamples, upfeats)]), dim=1)
-
-        return self.block(x)
-
-
-class UNet3p(nn.Module):
-    """Implements a UNet3+ architecture
-
-    Args:
-        layout (list<int>): number of channels after each contracting block
-        in_channels (int, optional): number of channels in the input tensor
-        num_classes (int, optional): number of output classes
-        act_layer: activation layer
-        norm_layer: normalization layer
-        drop_layer: dropout layer
-        conv_layer: convolutional layer
-    """
-    def __init__(
-        self,
-        layout: List[int],
-        in_channels: int = 3,
-        num_classes: int = 10,
-        act_layer: Optional[nn.Module] = None,
-        norm_layer: Optional[Callable[[int], nn.Module]] = None,
-        drop_layer: Optional[Callable[..., nn.Module]] = None,
-        conv_layer: Optional[Callable[..., nn.Module]] = None
-    ) -> None:
-        super().__init__()
-
-        if act_layer is None:
-            act_layer = nn.ReLU(inplace=True)
-        if norm_layer is None:
-            norm_layer = nn.BatchNorm2d
-
-        # Contracting path
-        self.encoders = nn.ModuleList([])
-        _layout = [in_channels] + layout
-        _pool = False
-        for in_chan, out_chan in zip(_layout[:-1], _layout[1:]):
-            self.encoders.append(down_path(in_chan, out_chan, _pool, 1, act_layer, norm_layer, drop_layer, conv_layer))
-            _pool = True
-
-        # Expansive path
-        self.decoders = nn.ModuleList([])
-        for row in range(len(layout) - 1):
-            self.decoders.append(FSAggreg(layout[:row],
-                                          layout[row],
-                                          [len(layout) * layout[0]] * (len(layout) - 2 - row) + layout[-1:],
-                                          act_layer, norm_layer, drop_layer, conv_layer))
-
-        # Classifier
-        self.classifier = nn.Conv2d(len(layout) * layout[0], num_classes, 1)
-
-        init_module(self, 'relu')
-
-    def forward(self, x: Tensor) -> Tensor:
-
-        xs: List[Tensor] = []
-        # Contracting path
-        for encoder in self.encoders:
-            xs.append(encoder(xs[-1] if len(xs) > 0 else x))
-
-        # Full-scale expansive path
-        for idx in range(len(self.decoders) - 1, -1, -1):
-            xs[idx] = self.decoders[idx](xs[:idx], xs[idx], xs[idx + 1:])
-
-        # Classifier
-        x = self.classifier(xs[0])
         return x
 
 
@@ -446,57 +186,3 @@ def unet(pretrained: bool = False, progress: bool = True, **kwargs: Any) -> UNet
     """
 
     return _unet('unet', pretrained, progress, **kwargs)  # type: ignore[return-value]
-
-
-def unetp(pretrained: bool = False, progress: bool = True, **kwargs: Any) -> UNetp:
-    """UNet+ from `"UNet++: Redesigning Skip Connections to Exploit Multiscale Features in Image Segmentation"
-    <https://arxiv.org/pdf/1912.05074.pdf>`_
-
-    .. image:: https://github.com/frgfm/Holocron/releases/download/v0.1.3/unetp.png
-        :align: center
-
-    Args:
-        pretrained (bool): If True, returns a model pre-trained on ImageNet
-        progress (bool): If True, displays a progress bar of the download to stderr
-
-    Returns:
-        torch.nn.Module: semantic segmentation model
-    """
-
-    return _unet('unetp', pretrained, progress, **kwargs)  # type: ignore[return-value]
-
-
-def unetpp(pretrained: bool = False, progress: bool = True, **kwargs: Any) -> UNetpp:
-    """UNet++ from `"UNet++: Redesigning Skip Connections to Exploit Multiscale Features in Image Segmentation"
-    <https://arxiv.org/pdf/1912.05074.pdf>`_
-
-    .. image:: https://github.com/frgfm/Holocron/releases/download/v0.1.3/unetpp.png
-        :align: center
-
-    Args:
-        pretrained (bool): If True, returns a model pre-trained on ImageNet
-        progress (bool): If True, displays a progress bar of the download to stderr
-
-    Returns:
-        torch.nn.Module: semantic segmentation model
-    """
-
-    return _unet('unetpp', pretrained, progress, **kwargs)  # type: ignore[return-value]
-
-
-def unet3p(pretrained: bool = False, progress: bool = True, **kwargs: Any) -> UNet3p:
-    """UNet3+ from
-    `"UNet 3+: A Full-Scale Connected UNet For Medical Image Segmentation" <https://arxiv.org/pdf/2004.08790.pdf>`_
-
-    .. image:: https://github.com/frgfm/Holocron/releases/download/v0.1.3/unet3p.png
-        :align: center
-
-    Args:
-        pretrained (bool): If True, returns a model pre-trained on ImageNet
-        progress (bool): If True, displays a progress bar of the download to stderr
-
-    Returns:
-        torch.nn.Module: semantic segmentation model
-    """
-
-    return _unet('unet3p', pretrained, progress, **kwargs)  # type: ignore[return-value]
